requirements.txt
aiofiles==23.1.0
aiohttp==3.8.6
aiosignal==1.3.1
Async-OAuthlib==0.0.9
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.8.2
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.6
discord.py==2.3.2
dnspython==2.4.1
Flask==3.0.3
frozenlist==1.4.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.14.4
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
libsass==0.22.0
Markdown==3.5.1
MarkupSafe==2.1.3
motor==3.2.0
multidict==6.0.4
oauthlib==3.2.2
Pillow==10.1.0
priority==2.0.0
PyJWT==2.8.0
pymongo==4.4.1
python-dotenv==1.0.0
Quart==0.19.6
Quart-Discord==2.1.4
Quart-Motor==2.4.7
quart-rate-limiter==0.10.0
requests==2.31.0
six==1.16.0
soupsieve==2.5
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.0.7
Werkzeug==3.0.3
wsproto==1.2.0
yarl==1.9.2
zipp==3.16.2
